+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/data/db/Db_Fixtures.kt
@@ -1,4 +1,4 @@
-package test.db
+package test.data.db
 
 import kiit.common.DateTimes
 import org.threeten.bp.LocalDate
