+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/data/repo/Repo_Common_Tests.kt
@@ -1,4 +1,4 @@
-package test.entities.repo
+package test.data.repo
 
 import kiit.common.utils.Random
 import kiit.entities.EntityRepo
